refactor: extract parallel test helpers
